feat: move CollectionInputLevel to its own file
