+++ src/collection/domain/models/Collection.ts
@@ -1,6 +1,7 @@
 import { UpwardHierarchyNode } from '../../../shared/hierarchy/domain/models/UpwardHierarchyNode'
 import { CollectionContact } from './CollectionContact'
 import { CollectionType } from './CollectionType'
+import { CollectionInputLevel } from './CollectionInputLevel'
 
 export interface Collection {
   id: string
@@ -14,10 +15,4 @@ export interface Collection {
   contacts: CollectionContact[]
   isMetadataBlockRoot: boolean
   isFacetRoot: boolean
-}
-
-export interface CollectionInputLevel {
-  datasetFieldName: string
-  include: boolean
-  required: boolean
 }
