+++ src/collection/domain/models/CollectionInputLevel.ts
@@ -1,0 +1,5 @@
+export interface CollectionInputLevel {
+  datasetFieldName: string
+  include: boolean
+  required: boolean
+}
